docs : added README.md with my first project description and details
The Tic Tac Toe game is a simple web-based project built using HTML, CSS, and JavaScript. It allows two players to play the classic X-and-O game on a 3x3 grid. The game includeturn switching, win detection, draw detection, and a reset button. It demonstrates basic DOM manipulation,event handling, and logical game state management. This project was created as part of a daily coding challenge to practice frontend skills and build small but complete applications.